fix: 빌드 오류 해결 - 클라이언트 컴포넌트, 서버 컴포넌트 간의 문제
